Allow user overrriding of fonts.
GuiFont.ini may be in the user's configuration directory and may reference a
font by name if it is in there too or in data/fonts or by full path.  The
".ttf" suffix may be omitted; it will be appended automatically iff there
are no dots present in the base name.

GuiFont.ini may be in the user's configuration directory.